Split out timing code
