+++ src/plugin.rs
@@ -1,6 +1,6 @@
 //! Contains main plugin exported by this crate.
 
-use crate::action_state::{ActionData, ActionState, Timing};
+use crate::action_state::{ActionData, ActionState};
 use crate::axislike::{
     AxisType, DeadZoneShape, DualAxis, DualAxisData, MouseMotionAxisType, MouseWheelAxisType,
     SingleAxis, VirtualAxis, VirtualDPad,
@@ -8,6 +8,7 @@ use crate::axislike::{
 use crate::buttonlike::{MouseMotionDirection, MouseWheelDirection};
 use crate::clashing_inputs::ClashStrategy;
 use crate::input_map::InputMap;
+use crate::timing::Timing;
 use crate::user_input::{InputKind, Modifier, UserInput};
 use crate::Actionlike;
 use core::hash::Hash;
